Move Config validations into it's own module
